requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
pkg-resources==0.0.0
PyAudio==0.2.11
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SpeechRecognition==3.8.1
tornado==6.1
tzdata==2022.1
tzlocal==4.2
